refactor: добавит общую папку для форм точек
