+++ src/view/point-form/edit-point-destinations.js
