+++ src/view/point-form/edit-point-helpers.js
