+++ src/view/point-form/edit-point-offers.js
