+++ src/view/point-form/edit-point.js
@@ -1,13 +1,13 @@
 import dayjs from 'dayjs';
-import { getFirstLetterInCapitalLetters } from '../utils/common';
+import { getFirstLetterInCapitalLetters } from '../../utils/common';
 import flatpickr from 'flatpickr';
-import SmartView from './smart';
-import { formValidity } from '../utils/form-validity';
-import { BLANK_POINT } from '../const';
+import SmartView from '../smart';
+import { formValidity } from '../../utils/form-validity';
+import { BLANK_POINT } from '../../const';
 import { createAdditionalOffer } from './edit-point-offers';
 import { createDestinationTemplate } from './edit-point-destinations';
 
-import '../../node_modules/flatpickr/dist/flatpickr.min.css';
+import '../../../node_modules/flatpickr/dist//flatpickr.min.css';
 import { createCityList, createContentButton, createEventRollupButtonTemplate, createEventTypeList } from './edit-point-helpers';
 
 const createEditPointForm = (data, AllOffers, Alldestinations, isNewEvent) => {
